add support for checking all of a user's repos
if no reposlug is given a list of all repos for the authenticated user is
retrieved and subsequently checked for unreleased commits. There are two
options to control what is included in the output. --forks will include forks
in the output, whereas --no-tags shows repositories without a single release
as well.